# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Real-Time Event Streaming Platform API
  version: 0.1.0
  description: |
    Planned REST contract for asynchronous order-event ingestion and processed
    event queries. The repository may not yet implement every documented
    operation. A successful ingestion response confirms acceptance, not
    completion of downstream Kafka processing.
  contact:
    name: Project Maintainer
tags:
  - name: Ingestion
    description: Submit events for asynchronous processing.
  - name: Events
    description: Query processed events.
  - name: Statistics
    description: Query aggregate processing statistics.
paths:
  /api/v1/events:
    post:
      tags:
        - Ingestion
      servers:
        - url: http://localhost:8080
          description: Local ingestion service
      summary: Submit an order event
      description: |
        Validates and accepts an event for asynchronous publication to Kafka.
        HTTP 202 does not mean that the Processing Service has completed the
        event.
      operationId: submitOrderEvent
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderEventRequest'
            examples:
              orderCreated:
                summary: Order created
                value:
                  eventType: ORDER_CREATED
                  eventVersion: 1
                  orderId: order-10001
                  occurredAt: '2026-07-25T18:30:00Z'
                  source: demo-order-service
                  payload:
                    customerReference: customer-200
                    currency: USD
                    totalAmount: 149.99
                    itemCount: 2
      responses:
        '202':
          description: Event accepted for asynchronous processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: Request payload is too large.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: Request rate limit exceeded, if rate limiting is enabled.
          headers:
            Retry-After:
              description: Suggested delay in seconds before retrying.
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          description: Event ingestion is temporarily unavailable.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags:
        - Events
      servers:
        - url: http://localhost:8081
          description: Local processing/query service
      summary: List processed events
      operationId: listEvents
      parameters:
        - name: orderId
          in: query
          description: Return events for one order.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: eventType
          in: query
          description: Return events of one type.
          schema:
            $ref: '#/components/schemas/OrderEventType'
        - name: status
          in: query
          description: Return events with one processing status.
          schema:
            $ref: '#/components/schemas/ProcessingStatus'
        - name: from
          in: query
          description: Inclusive lower bound for event occurrence time.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Exclusive upper bound for event occurrence time.
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Zero-based page number.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of records per page.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort expression. Initially restricted to documented fields.
          schema:
            type: string
            default: occurredAt,desc
            example: occurredAt,desc
      responses:
        '200':
          description: A page of processed events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/events/{eventId}:
    get:
      tags:
        - Events
      servers:
        - url: http://localhost:8081
          description: Local processing/query service
      summary: Get a processed event
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Processed event details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessedEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/statistics:
    get:
      tags:
        - Statistics
      servers:
        - url: http://localhost:8081
          description: Local processing/query service
      summary: Get event-processing statistics
      operationId: getProcessingStatistics
      parameters:
        - name: from
          in: query
          description: Inclusive beginning of the statistics window.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Exclusive end of the statistics window.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Aggregate event-processing statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatistics'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      description: Unique event identifier.
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Client correlation identifier. The service generates one if absent.
      schema:
        type: string
        minLength: 1
        maxLength: 100
  responses:
    BadRequest:
      description: The request is malformed or violates validation rules.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: An unexpected internal error occurred.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    CreateOrderEventRequest:
      type: object
      additionalProperties: false
      required:
        - eventType
        - eventVersion
        - orderId
        - occurredAt
        - source
        - payload
      properties:
        eventType:
          $ref: '#/components/schemas/OrderEventType'
        eventVersion:
          type: integer
          minimum: 1
          default: 1
        orderId:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
        occurredAt:
          type: string
          format: date-time
        source:
          type: string
          minLength: 1
          maxLength: 100
        payload:
          description: Event-specific data matching `eventType`.
          oneOf:
            - $ref: '#/components/schemas/OrderCreatedPayload'
            - $ref: '#/components/schemas/PaymentCompletedPayload'
            - $ref: '#/components/schemas/PaymentFailedPayload'
            - $ref: '#/components/schemas/OrderCancelledPayload'
            - $ref: '#/components/schemas/OrderShippedPayload'
    EventAcceptedResponse:
      type: object
      additionalProperties: false
      required:
        - eventId
        - correlationId
        - status
        - acceptedAt
      properties:
        eventId:
          type: string
          format: uuid
        correlationId:
          type: string
        status:
          type: string
          enum:
            - ACCEPTED
        acceptedAt:
          type: string
          format: date-time
        statusUrl:
          type: string
          format: uri
          description: URL that can be queried for eventual processing status.
    ProcessedEvent:
      type: object
      additionalProperties: false
      required:
        - eventId
        - eventType
        - eventVersion
        - orderId
        - occurredAt
        - receivedAt
        - source
        - correlationId
        - status
        - payload
      properties:
        eventId:
          type: string
          format: uuid
        eventType:
          $ref: '#/components/schemas/OrderEventType'
        eventVersion:
          type: integer
          minimum: 1
        orderId:
          type: string
        occurredAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
        correlationId:
          type: string
        status:
          $ref: '#/components/schemas/ProcessingStatus'
        payload:
          description: Event-specific data matching `eventType`.
          oneOf:
            - $ref: '#/components/schemas/OrderCreatedPayload'
            - $ref: '#/components/schemas/PaymentCompletedPayload'
            - $ref: '#/components/schemas/PaymentFailedPayload'
            - $ref: '#/components/schemas/OrderCancelledPayload'
            - $ref: '#/components/schemas/OrderShippedPayload'
        failure:
          allOf:
            - $ref: '#/components/schemas/FailureSummary'
          nullable: true
    FailureSummary:
      type: object
      additionalProperties: false
      properties:
        category:
          type: string
          description: Safe, non-sensitive failure category.
        message:
          type: string
          description: Safe failure summary without a stack trace.
        failedAt:
          type: string
          format: date-time
        retryCount:
          type: integer
          minimum: 0
    EventPage:
      type: object
      additionalProperties: false
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProcessedEvent'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    ProcessingStatistics:
      type: object
      additionalProperties: false
      required:
        - windowStart
        - windowEnd
        - totalEvents
        - processedEvents
        - failedEvents
        - eventsByType
      properties:
        windowStart:
          type: string
          format: date-time
        windowEnd:
          type: string
          format: date-time
        totalEvents:
          type: integer
          format: int64
          minimum: 0
        processedEvents:
          type: integer
          format: int64
          minimum: 0
        failedEvents:
          type: integer
          format: int64
          minimum: 0
        averageProcessingLatencyMs:
          type: number
          format: double
          minimum: 0
          nullable: true
        eventsByType:
          type: object
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
    OrderEventType:
      type: string
      enum:
        - ORDER_CREATED
        - PAYMENT_COMPLETED
        - PAYMENT_FAILED
        - ORDER_CANCELLED
        - ORDER_SHIPPED
    OrderCreatedPayload:
      type: object
      additionalProperties: true
      required:
        - customerReference
        - currency
        - totalAmount
        - itemCount
      properties:
        customerReference:
          $ref: '#/components/schemas/SafeIdentifier'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        totalAmount:
          $ref: '#/components/schemas/PositiveMoney'
        itemCount:
          type: integer
          minimum: 1
    PaymentCompletedPayload:
      type: object
      additionalProperties: true
      required:
        - paymentReference
        - amount
        - currency
        - completedAt
      properties:
        paymentReference:
          $ref: '#/components/schemas/SafeIdentifier'
        amount:
          $ref: '#/components/schemas/PositiveMoney'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        completedAt:
          type: string
          format: date-time
    PaymentFailedPayload:
      type: object
      additionalProperties: true
      required:
        - paymentReference
        - failureReason
        - failedAt
      properties:
        paymentReference:
          $ref: '#/components/schemas/SafeIdentifier'
        failureReason:
          $ref: '#/components/schemas/PaymentFailureReason'
        failedAt:
          type: string
          format: date-time
    OrderCancelledPayload:
      type: object
      additionalProperties: true
      required:
        - reason
        - cancelledAt
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 250
        cancelledAt:
          type: string
          format: date-time
    OrderShippedPayload:
      type: object
      additionalProperties: true
      required:
        - shipmentReference
        - carrier
        - shippedAt
      properties:
        shipmentReference:
          $ref: '#/components/schemas/SafeIdentifier'
        carrier:
          type: string
          minLength: 1
          maxLength: 100
        shippedAt:
          type: string
          format: date-time
    PaymentFailureReason:
      type: string
      enum:
        - DECLINED
        - INSUFFICIENT_FUNDS
        - EXPIRED_PAYMENT_METHOD
        - PROCESSOR_UNAVAILABLE
        - VALIDATION_ERROR
        - UNKNOWN
    SafeIdentifier:
      type: string
      minLength: 1
      maxLength: 100
      pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      example: USD
    PositiveMoney:
      type: number
      minimum: 0
      exclusiveMinimum: true
      multipleOf: 0.01
      example: 149.99
    ProcessingStatus:
      type: string
      enum:
        - RECEIVED
        - PROCESSING
        - PROCESSED
        - FAILED
        - DEAD_LETTERED
    ApiError:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - path
      properties:
        type:
          type: string
          format: uri
          description: Stable identifier for the error category.
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        correlationId:
          type: string
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidationError'
    FieldValidationError:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string